Upd:pass conversions

+++ exercises/conversions/using_as.rs
@@ -1,7 +1,7 @@
 // using_as.rs
 //
 // Type casting in Rust is done via the usage of the `as` operator. Please note
-// that the `as` operator is not only used when type casting. It also helps with
+// tator is not only used when type casting. Ithat the `as` oper also helps with
 // renaming imports.
 //
 // The goal is to make sure that the division does not fail to compile and
@@ -10,11 +10,9 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
